refactor chat components into their own folder
