+++ components/Chat/ChatContainer.jsx
@@ -5,11 +5,12 @@ import {
   View,
   TextInput,
   TouchableOpacity,
-  Text,
-  StyleSheet
+  Text
 } from "react-native";
 
-const ChatContainer = ({ socket, chatLog, setChatLog }) => {
+import styles from "./ChatContainerStyles";
+
+const ChatContainer = ({ socket, chatLog }) => {
   const [chatText, setChatText] = useState("");
 
   return (
@@ -39,39 +40,6 @@ const ChatContainer = ({ socket, chatLog, setChatLog }) => {
   );
 };
 
-const styles = StyleSheet.create({
-  chatContainer: {
-    height: "40%",
-    width: "100%"
-  },
 
-  chatSendContainer: {
-    height: "10%",
-    width: "100%",
-    flexDirection: "row"
-  },
-
-  chatBox: {
-    backgroundColor: "yellow",
-    width: "80%"
-  },
-
-  chatLog: {
-    height: "60%",
-    width: "100%",
-    backgroundColor: "white",
-    borderColor: "black",
-    padding: 3
-  },
-
-  sendMessageButton: {
-    backgroundColor: "green",
-    alignItems: "center",
-    justifyContent: "center",
-    padding: 5,
-    height: "100%",
-    width: "20%"
-  }
-});
 
 export default ChatContainer;
